Added location checking on loading to prevent plugin errors and allows for doubles in setting TPC

+++ src/com/AustinPilz/FridayThe13th/Components/Arena/Arena.java
@@ -158,7 +158,7 @@ public class Arena
      *
      * @param value
      */
-    public void setMinutesPerCounselor(int value) {
+    public void setMinutesPerCounselor(double value) {
         minutesPerCounselor = value;
         updateInDB();
     }
